feat: add template for generating default app name
This commit adds a new template file `_helpers.tpl` which defines a function `chatwoot-bot.fullname`. This function generates a default fully qualified app name for the chatwoot-bot application. It takes into consideration the `fullnameOverride` value from the Helm chart values and truncates the name to 63 characters as per the DNS naming spec.

The purpose of this change is to provide a consistent and standardized way of generating the app name, ensuring compatibility with Kubernetes name fields limitations.

Note: This commit message is a suggestion and should be modified as per the repository's commit message conventions.